+++ app/lib/definitions.ts
@@ -1,0 +1,8 @@
+export type Screenings = {
+    id: number;
+    start_time: string;
+  };
+  
+export type SpecificScreeningsResponse = {
+    data: Screenings[];
+  };
