+++ app/page.tsx
@@ -1,5 +1,5 @@
 import MainNavbar from './ui/MainNavbar';
-import TestComponent from './components/testComponent';
+import TestComponent from './components/movie/testComponent';
 
 export default function Home() {
   return (
